remove function added
