+++ src/app/lib/actions.js
@@ -36,3 +36,56 @@ export async function createProject(status, project_name, date) {
         throw new Error("An error occured")
     }
 }
+
+
+export async function addMember(name, email, position, password, admin) {
+    console.log(name, email, position, password, admin)
+    try {
+        if (!name || !email || !position) {
+            throw new Error("Missing fields: name, email, or position");
+        }
+
+        const hashPassword = await bcrypt.hash(password, 10);
+
+        const data = await sql`
+            INSERT INTO users (name, email, position, admin, password)
+            VALUES (${name}, ${email}, ${position}, ${admin}, ${hashPassword})
+            RETURNING *;
+        `;
+        console.log("Added member successfully:", data.rows[0]);
+        return data.rows[0];
+    } catch (error) {
+        console.error("Error adding member:", error);
+        throw new Error("Error adding member");
+    }
+}
+
+export async function deleteMember(
+    prevState,
+    formData
+) {
+    try {
+        const parsedFormData = Object.fromEntries(formData.entries());
+
+        console.log(parsedFormData)
+        if (!parsedFormData.user_id) {
+            throw new Error("Missing required field: email");
+        }
+
+        const result = await sql`
+            DELETE FROM users
+            WHERE id = ${parsedFormData.user_id}
+            RETURNING *;
+        `;
+
+        if (result.rowCount === 0) {
+            throw new Error(`No member found with id: ${parsedFormData.user_id}`);
+        }
+
+        console.log("Deleted member successfully:", result.rows[0]);
+        return result.rows[0];
+    } catch (error) {
+        console.error("Error deleting member:", error);
+        throw new Error("Error deleting member");
+    }
+}
