+++ src/app/seed/route.js
@@ -116,14 +116,14 @@ export async function GET() {
     console.log("Seeding Guidelines...")
     await seedGuidelines()
 
-    console.log("Seeding Users...");
-    await seedUsers();
+    // console.log("Seeding Users...");
+    // await seedUsers();
 
-    console.log("Seeding Updates....");
-    await seedUpdates();
+    // console.log("Seeding Updates....");
+    // await seedUpdates();
 
-    console.log("Seeding Projects...");
-    await seedProjects();
+    // console.log("Seeding Projects...");
+    // await seedProjects();
 
     console.log("Before Commit");
     await client.sql`COMMIT`;
